complete connection to Clearblade obj
